+++ Asteroids-client/utils.h
@@ -1,0 +1,2 @@
+#pragma once
+constexpr const float DEG_TO_RAD = 3.14159f / 180.0f;
